Change file extensions .cc -> .h
