+++ pycolmap/estimators/absolute_pose.h
