+++ pycolmap/estimators/alignment.h
