+++ pycolmap/estimators/essential_matrix.h
