+++ pycolmap/estimators/generalized_absolute_pose.h
