+++ pycolmap/estimators/homography_matrix.h
