+++ pycolmap/estimators/triangulation.h
