+++ pycolmap/estimators/two_view_geometry.h
