+++ pycolmap/geometry/bindings.h
@@ -10,7 +10,7 @@
 #include <pybind11/stl.h>
 #include <pybind11/stl_bind.h>
 
-#include "pycolmap/geometry/homography_matrix.cc"
+#include "pycolmap/geometry/homography_matrix.h"
 
 namespace py = pybind11;
 using namespace pybind11::literals;
