+++ pycolmap/geometry/homography_matrix.h
