+++ pycolmap/optim/bindings.h
