+++ pycolmap/pipeline/extract_features.h
