+++ pycolmap/pipeline/images.h
