+++ pycolmap/pipeline/meshing.h
