+++ pycolmap/pipeline/sfm.h
