+++ pycolmap/scene/camera.h
