+++ pycolmap/scene/image.h
