+++ pycolmap/scene/point3D.h
