+++ pycolmap/scene/reconstruction.h
