+++ pycolmap/scene/track.h
