+++ pycolmap/sfm/incremental_triangulator.h
